Refactor the way git tasks are called and errors are captured.
