+++ apps/rush-lib/src/logic/policy/GitEmailPolicy.ts
@@ -4,74 +4,37 @@
 import * as colors from 'colors';
 import * as os from 'os';
 
-import { GitPolicy } from './GitPolicy';
 import { RushConfiguration } from '../../api/RushConfiguration';
 import { AlreadyReportedError } from '../../utilities/AlreadyReportedError';
 import { Utilities } from '../../utilities/Utilities';
+import { Git } from '../Git';
 
-interface IResultOrError<TResult> {
-  error?: Error;
-  result?: TResult;
-}
-
-export class GitEmailPolicy extends GitPolicy {
-  public static getUserEmail(rushConfiguration: RushConfiguration): string {
-    // Determine the user's account
-    // Ex: "bob@example.com"
-    const emailResult: IResultOrError<string> = GitEmailPolicy.tryGetEmail();
-    if (emailResult.error) {
-      console.log(
-        [
-          `Error: ${emailResult.error.message}`,
-          'Unable to determine your Git configuration using this command:',
-          '',
-          '    git config user.email',
-          '',
-          `If you didn't configure your e-mail yet, try something like this:`,
-          ...GitEmailPolicy.getEmailExampleLines(rushConfiguration)
-        ].join(os.EOL)
-      );
-      throw new AlreadyReportedError();
+export class GitEmailPolicy {
+  public static validate(rushConfiguration: RushConfiguration ): void {
+    const gitPath: string | undefined = Git.getGitPath();
+    if (!gitPath) {
+      // Git isn't present, so the git policy doesn't apply
+      return;
     }
 
-    return emailResult.result || '';
-  }
-
-  private static tryGetEmail(): IResultOrError<string> {
+    let userEmail: string;
     try {
-      return {
-        result: Utilities.executeCommandAndCaptureOutput(
-          'git',
-          ['config', 'user.email'],
-          '.'
-        ).trim()
-      };
+      userEmail = Git.getGitEmail(rushConfiguration);
     } catch (e) {
-      return {
-        error: e
-      };
-    }
-  }
-
-  private static getEmailExampleLines(rushConfiguration: RushConfiguration): string[] {
-    return [
-      colors.cyan('    git config --local user.name "Mr. Example"'),
-      colors.cyan(`    git config --local user.email "${rushConfiguration.gitSampleEmail || 'example@contoso.com'}"`)
-    ];
-  }
-
-  protected innerValidate(gitPath: string, rushConfiguration: RushConfiguration ): void {
-    const userEmail: string = GitEmailPolicy.getUserEmail(rushConfiguration);
-    if (!userEmail) {
-      console.log(colors.red('Aborting, so you can go fix your settings.  (Or use --bypass-policy to skip.)'));
-      throw new AlreadyReportedError();
+      if (e instanceof AlreadyReportedError) {
+        console.log(colors.red('Aborting, so you can go fix your settings.  (Or use --bypass-policy to skip.)'));
+        throw e;
+      } else {
+        console.log(colors.red(`An unexpected error occurred: ${e}`));
+        throw e;
+      }
     }
 
-        // sanity check; a valid e-mail should not contain any whitespace
+    console.log('Checking Git policy for this repository.' + os.EOL);
+
+    // sanity check; a valid e-mail should not contain any whitespace
     // if this fails, then we have another issue to report
-    if (userEmail && userEmail.match(/^\S+$/g)) {
-       console.log('Checking Git policy for this repository.' + os.EOL);
-
+    if (userEmail.match(/^\S+$/g)) {
       if (rushConfiguration.gitAllowedEmailRegExps.length === 0) {
         return;
       }
@@ -88,8 +51,11 @@ export class GitEmailPolicy extends GitPolicy {
     // Ex. "Mr. Example <mr@example.com>"
     let fancyEmail: string = colors.cyan(userEmail);
     try {
-      const userName: string = Utilities.executeCommandAndCaptureOutput('git',
-        ['config', 'user.name'], '.').trim();
+      const userName: string = Utilities.executeCommandAndCaptureOutput(
+        gitPath,
+        ['config', 'user.name'],
+        '.'
+      ).trim();
       if (userName) {
         fancyEmail = `${userName} <${fancyEmail}>`;
       }
@@ -97,7 +63,6 @@ export class GitEmailPolicy extends GitPolicy {
       // but if it fails, this isn't critical, so don't bother them about it
     }
 
-    debugger;
     console.log([
       'Hey there!  To keep things tidy, this repo asks you to submit your Git commits using an e-mail like ',
       rushConfiguration.gitAllowedEmailRegExps.length > 1 ? 'one of these patterns:' : 'this pattern:',
@@ -116,5 +81,12 @@ export class GitEmailPolicy extends GitPolicy {
 
     console.log(colors.red('Aborting, so you can go fix your settings.  (Or use --bypass-policy to skip.)'));
     throw new AlreadyReportedError();
+  }
+
+  public static getEmailExampleLines(rushConfiguration: RushConfiguration): string[] {
+    return [
+      colors.cyan('    git config --local user.name "Mr. Example"'),
+      colors.cyan(`    git config --local user.email "${rushConfiguration.gitSampleEmail || 'example@contoso.com'}"`)
+    ];
   }
 }
